Test: add separate config and split tests
