+++ src/test/resources/application-test.conf
@@ -1,0 +1,24 @@
+ktor {
+  application {
+    modules = [de.pause.ApplicationKt.module]
+  }
+
+  deployment {
+    port = 8080
+  }
+
+  datasource {
+    driver = "org.h2.Driver"
+    url = "jdbc:h2:mem:test;DB_CLOSE_DELAY=-1;"
+    user = "sa"
+    password = "pass"
+  }
+
+  jwt {
+    secret = "test_secret"
+    issuer = "test_issuer"
+    audience = "test_audience"
+    realm = "test_realm"
+    expiration = 60
+  }
+}
